requirements.txt
wheel
pycryptodome==3.9.8
tinyec==0.3.1
pymongo
flask
flask_api
requests
cryptography
